requirements.txt
# Core RAG Pipeline Dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
python-multipart==0.0.6

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5.1

# Vector Database and Embeddings
chromadb==0.4.15
langchain==0.0.335
langchain-community==0.0.6

# Content Generation
openai==1.3.5
requests==2.31.0

# File Generation
python-pptx==0.6.23
reportlab==4.0.7
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
pandas==2.1.4
numpy==1.25.2
Pillow==10.1.0
jinja2==3.1.2
pydantic==2.5.0
